Реорганизация компонентов роутинга
